feat: split cobra functions
